PATCH -> 사장님이 주문 확정한 상품 개수만큼 DB Item개수 차감 기능 추가

+++ src/main/java/zupzup/back_end/reservation/domain/type/OrderSpecific.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Embeddable
 public class OrderSpecific {
     //[상품 이름, 가격, 갯수, (이미지)]
-    private int itemId; // 일단 단순 int로 받아오고, 외래키로 갈지 말지는 나중에
+    private Long itemId; // 일단 단순 Long으로 받아오고, 외래키로 갈지 말지는 나중에
     private String itemName;    // 상품 이름
     private int itemPrice;  // 가격
     @Min(value = 0, message = "상품의 개수는 0개 이하일 수 없습니다.")
